chore(csharp): update csharp sdk to latest contracts (#1839)
Update all tcp and http contracts to latest version, tests and examples.
Also small cleanup

+++ foreign/csharp/Iggy_SDK/Contracts/Http/TopicRequest.cs
@@ -23,7 +23,7 @@ public record TopicRequest(
     int? TopicId, // If not provided, the Iggy server will generate one automatically
     string Name,
     CompressionAlgorithm CompressionAlgorithm,
-    int MessageExpiry,
+    ulong MessageExpiry,
     int PartitionsCount,
     byte ReplicationFactor,
     ulong MaxTopicSize);
